requirements.txt
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-ldap==3.4.3
